pyproject.toml
[project]
name = "airflow-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # MLOps Core Libraries
    "mlflow>=2.8.0",
    "great-expectations>=0.18.0",
    "dvc>=3.0.0",
    # ML Libraries
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Data Processing
    "pyarrow>=12.0.0",
    "fastparquet>=0.8.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    # Model Serving
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    # Monitoring & Observability
    "prometheus-client>=0.17.0",
    "wandb>=0.15.0",
    # Experiment Tracking
    "optuna>=3.3.0",
    "hyperopt>=0.2.7",
    "pillow>=11.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "papermill>=2.4.0",
]
